separate sponsor file
